Move navigating or creating logic to report actions
